feat: 完成查看用户详情和查看个人信息接口

+++ api/internal/handler/user/setselfinfohandler.go
@@ -20,7 +20,7 @@ func SetSelfInfoHandler(svcCtx *svc.ServiceContext) http.HandlerFunc {
 			response.ErrWithMessage(r.Context(), w, "参数错误")
 			return
 		}
-		req.ID = utils.GetUserId(r)
+		req.ID = uint64(utils.GetUserId(r))
 		l := user.NewSetSelfInfoLogic(r.Context(), svcCtx)
 		_, err := l.SetSelfInfo(&req)
 		if err != nil {
