+++ api/internal/logic/user/setselfinfologic.go
@@ -25,7 +25,7 @@ func NewSetSelfInfoLogic(ctx context.Context, svcCtx *svc.ServiceContext) *SetSe
 
 func (l *SetSelfInfoLogic) SetSelfInfo(req *types.SetSelfInfoReq) (resp *types.SetSelfInfoResp, err error) {
 	_, err = l.svcCtx.User.SetSelfInfo(l.ctx, &user.SetSelfInfoReq{
-		Id:       req.ID,
+		Id:       int64(req.ID),
 		Username: req.Username,
 		Email:    req.Email,
 		Phone:    req.Phone,
